source_code/app/web/honeypot/requirements.txt
BeautifulSoup==3.2.2
certifi==2020.4.5.1
cffi==1.7.0
chardet==3.0.4
EditorConfig==0.12.2
filemagic==1.6
future==0.18.2
html5lib==1.0.1
HTMLParser==0.0.2
idna==2.9
jsbeautifier==1.11.0
logging==0.4.9.6
lxml==4.5.0
mechanize==0.4.5
normalize==2.0.2
pycparser==2.20
python-magic==0.4.15
requests==2.23.0
requests-file==1.4.3
richenum==1.3.1
six==1.14.0
tldextract==2.2.2
urllib3==1.25.8
webencodings==0.5.1
